Split schoology functions into separate files
